Init Conference Plan in code and init DB data

+++ src/main/java/com/example/conferencemanagmentapp/model/Lecture.java
@@ -1,18 +1,17 @@
 package com.example.conferencemanagmentapp.model;
 
 import lombok.Data;
-
 import java.time.LocalTime;
 import java.util.Map;
 
 @Data
 public class Lecture {
-    private Long id;
+    private int id;
     private LocalTime start;
     private LocalTime end;
     private Map<Integer, LectureRoot> lectureRoots;
 
-    public Lecture(Long id, LocalTime start, LocalTime end, Map<Integer, LectureRoot> lectureRoots) {
+    public Lecture(int id, LocalTime start, LocalTime end, Map<Integer, LectureRoot> lectureRoots) {
         this.id = id;
         this.start = start;
         this.end = end;
